# Introduce lighthouse-ci performance assertions