--- a/lighthouserc.js
+++ b/lighthouserc.js
@@ -7,6 +7,7 @@ module.exports = {
       puppeteerScript: './scripts/lighthouse/puppeteer-script.js',
       settings: {
         onlyCategories: "accessibility,best-practices,performance,seo",
+        disableStorageReset: true
       }
     },
     upload: {