--- a/lighthouserc.js
+++ b/lighthouserc.js
@@ -3,7 +3,7 @@ module.exports = {
     collect: {
       url: ['http://localhost:9000/Article?url=https://www.theguardian.com/commentisfree/2020/feb/08/hungary-now-for-the-new-right-what-venezuela-once-was-for-the-left#noads'],
       startServerCommand: 'NODE_ENV=production DISABLE_LOGGING_AND_METRICS=true node dist/frontend.server.js',
-      numberOfRuns: '6',
+      numberOfRuns: '10',
       puppeteerScript: './scripts/lighthouse/puppeteer-script.js',
       settings: {
         onlyCategories: "accessibility,best-practices,performance,seo",