--- a/lighthouserc.js
+++ b/lighthouserc.js
@@ -13,5 +13,14 @@ module.exports = {
     upload: {
       target: 'temporary-public-storage',
     },
+    assert: {
+      includePassedAssertions: true,
+      assertions: {
+        "first-contentful-paint": ["warn", {"maxNumericValue": 1000}],
+        "largest-contentful-paint": ["warn", {"maxNumericValue": 3000}],
+        "interactive": ["warn", {"maxNumericValue": 3500}],
+        "cumulative-layout-shift": ["warn", {"maxNumericValue": 0.001}],
+      }
+    }
   },
 };