@PRBuilds 2020-11-12
PRbuilds results:










&#x1F49A; **AMP validation**
[amp-report.txt](https://s3-eu-west-1.amazonaws.com/prbuilds/PR-2042/amp-report.txt)







**LightHouse Reporting**



--automated message