@jorgeazevedo 2020-11-13
@oliverlloyd thanks! So this will make the status check fail, but lighthouse is not a "required check" so it won't block merging.